# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aaryajanani Content Ops & Project Memory Platform API
  version: 1.0.0
  description: |
    Backend API for the Content Operations & Project Memory Platform (Aaryajanani).
    Node.js / Express / MongoDB.

    ## Authentication
    - Register or login to obtain an `accessToken` (JWT, 15 min) and a `refreshToken`.
    - Send the access token as `Authorization: Bearer <token>` on protected endpoints.
    - When the access token expires, call `POST /auth/refresh` with the `refreshToken`
      (rotates the token; the old refresh token is revoked).
    - Logout revokes the current refresh token.
  contact:
    name: Aaryajanani Backend
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Local development (default apiPrefix /api/v1)
...
paths:
...
  /projects/{projectId}/tasks:
    get:
      tags: [Tasks]
      summary: List project tasks (kanban columns filterable)
      operationId: listTasks
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/TaskStatus'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Skip'
      responses:
        '200':
          description: Paginated tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Task'
                      total:
                        type: integer
                      limit:
                        type: integer
                      skip:
                        type: integer
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Tasks]
      summary: Create a task in a project
      operationId: createTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing task.create permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /projects/{projectId}/tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: Get a single task
      operationId: getTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task found
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Tasks]
      summary: Update a task (title, priority, assignee, due date)
      operationId: updateTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing task.update permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Tasks]
      summary: Delete a task
      operationId: deleteTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      deleted:
                        type: boolean
                        example: true
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/{projectId}/tasks/{taskId}/status:
    post:
      tags: [Tasks]
      summary: Move a task between kanban columns
      operationId: setTaskStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskStatusRequest'
      responses:
        '200':
          description: Task status updated (assignee + managers notified)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ================= Briefs =================
...
  /projects/{projectId}/reviews/summarize:
    post:
      tags: [Reviews]
      summary: Summarize unresolved feedback (rule-based, AI-ready)
      operationId: summarizeReview
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewSummarizeRequest'
      responses:
        '200':
          description: Action-oriented review summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ReviewSummary'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project or version not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/{projectId}/reviews/lock:
    post:
      tags: [Reviews]
      summary: Lock review feedback into the next revision scope
      operationId: lockReview
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewLockRequest'
      responses:
        '201':
          description: Feedback locked; revision request created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      revision:
                        $ref: '#/components/schemas/RevisionRequest'
                      lockedCommentCount:
                        type: integer
                      scopeItems:
                        type: array
                        items:
                          $ref: '#/components/schemas/ScopeItem'
        '400':
          description: No unresolved feedback for the version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing project.revision permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Project or version not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ================= Chat =================
...
  /organizations/{organizationId}/invoices:
    get:
      tags: [Invoices]
      summary: List invoices
      operationId: listInvoices
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [draft, sent, paid, overdue, void]
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Skip'
      responses:
        '200':
          description: Paginated invoices
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Invoice'
                      total:
                        type: integer
                      limit:
                        type: integer
                      skip:
                        type: integer
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing invoice.manage permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Invoices]
      summary: Create an invoice (draft)
      operationId: createInvoice
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceRequest'
      responses:
        '201':
          description: Invoice created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Invoice'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing invoice.manage permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /organizations/{organizationId}/invoices/{invoiceId}:
    get:
      tags: [Invoices]
      summary: Get an invoice
      operationId: getInvoice
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/InvoiceId'
      responses:
        '200':
          description: Invoice found
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Invoice'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags: [Invoices]
      summary: Edit a draft invoice
      operationId: updateInvoice
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/InvoiceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceRequest'
      responses:
        '200':
          description: Invoice updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Invoice'
        '400':
          description: Only draft invoices can be edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /organizations/{organizationId}/invoices/{invoiceId}/send:
    post:
      tags: [Invoices]
      summary: Send an invoice (draft -> sent, net-30 default)
      operationId: sendInvoice
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/InvoiceId'
      responses:
        '200':
          description: Invoice sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Invoice'
        '400':
          description: Only draft invoices can be sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /organizations/{organizationId}/invoices/{invoiceId}/void:
    post:
      tags: [Invoices]
      summary: Void an unpaid invoice
      operationId: voidInvoice
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationId'
        - $ref: '#/components/parameters/InvoiceId'
      responses:
        '200':
          description: Invoice voided
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Invoice'
        '400':
          description: Paid invoices cannot be voided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token returned by /auth/login or /auth/refresh (15 min expiry)
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
      description: MongoDB ObjectId of the organization
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
      description: MongoDB ObjectId of the project
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: Maximum number of items to return
    Skip:
      name: skip
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of items to skip (offset pagination)
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
      description: MongoDB ObjectId of the task
    TaskStatus:
      name: status
      in: query
      required: false
      schema:
        type: string
        enum: [todo, in_progress, in_review, done]
      description: Filter tasks by kanban column
...
    InvoiceId:
      name: invoiceId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'
      description: MongoDB ObjectId of the invoice
...
  schemas:
...
    RevisionRequest:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        sourceVersionId:
          type: string
          nullable: true
        requestedBy:
          type: string
          description: User id
        reason:
          type: string
          maxLength: 5000
        status:
          type: string
          enum: [in_progress, submitted, resolved]
        resolvedVersionId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              examples: [not_found, unauthorized, forbidden, conflict, internal_error]
            message:
              type: string
    ValidationError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              const: validation_error
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

    # ---------- Tasks ----------
    CreateTaskRequest:
      type: object
      required: [title]
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: 'Export 4K master'
        description:
          type: string
          maxLength: 2000
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        assignee:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
          description: User id responsible for the task
        dueDate:
          type: string
          format: date
    UpdateTaskRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        priority:
          type: string
          enum: [low, medium, high]
        assignee:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
        dueDate:
          type: string
          format: date
    TaskStatusRequest:
      type: object
      required: [status]
      additionalProperties: false
      properties:
        status:
          type: string
          enum: [todo, in_progress, in_review, done]
          example: done
    Task:
      type: object
      properties:
        _id:
          type: string
          example: 507f1f77bcf86cd799439011
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [todo, in_progress, in_review, done]
        priority:
          type: string
          enum: [low, medium, high]
        assignee:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        createdBy:
          type: string
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ---------- Briefs ----------
...
    ReviewSummarizeRequest:
      type: object
      required: [versionId]
      additionalProperties: false
      properties:
        versionId:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
          description: Version whose unresolved comments are summarized
    ReviewLockRequest:
      type: object
      required: [versionId]
      additionalProperties: false
      properties:
        versionId:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
          description: Version whose unresolved comments get locked into scope
    ReviewSummary:
      type: object
      properties:
        projectId:
          type: string
        versionId:
          type: string
        totalComments:
          type: integer
        unresolvedComments:
          type: integer
        actionItems:
          type: array
          items:
            type: string
          description: Action-oriented items extracted from feedback
        sections:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              items:
                type: array
                items:
                  type: string
        generatedBy:
          type: string
          const: rule-based-summarizer
          description: Swap point for an AI/LLM summarizer later
    ScopeItem:
      type: object
      properties:
        commentId:
          type: string
        author:
          type: string
        body:
          type: string

    # ---------- Chat ----------
...
    InvoiceRequest:
      type: object
      required: [number, amount]
      additionalProperties: false
      properties:
        projectId:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
        number:
          type: string
          minLength: 1
          maxLength: 100
          example: INV-2025-001
        amount:
          type: number
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
          default: USD
        dueDate:
          type: string
          format: date
        note:
          type: string
          maxLength: 2000
...
    Invoice:
      type: object
      properties:
        _id:
          type: string
        organizationId:
          type: string
        projectId:
          type: string
          nullable: true
        number:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [draft, sent, paid, overdue, void]
        issuedAt:
          type: string
          format: date-time
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
        paymentMethod:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...